issue38:: code refactor:: rewriting to ts.tsx/folder for tests/remove not used components
