+++ package.json
@@ -43,6 +43,7 @@
     "@types/express": "^4.17.9",
     "@types/express-urlrewrite": "^1.2.1",
     "@types/react-dom": "^16.9.8",
+    "@types/react-router-dom": "^5.1.7",
     "chart.js": "^2.9.4",
     "cypress": "^6.0.1",
     "express-urlrewrite": "^1.3.0",
