+++ src/components/pages/StatedGraphForm.tsx
@@ -1,11 +1,12 @@
 import React from "react";
-import Chart from "../Chart";
+
+import ChartFunctional from "../ChartFunctional";
 let GraphForm = () => {
   return (
     <div>
       <h1>GraphPage</h1>
-      <Chart
-        values={[
+      <ChartFunctional
+          values={[
           [new Date("December 17, 2020 03:24:00"), 20.2],
           [new Date("December 17, 2020 06:24:00"), 20.2],
           [new Date("December 19, 2020 12:24:00"), 42.2],
