+++ src/components/pages/StatedNameForm.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import StatedBalanceForm from "../StatedBalanceForm";
-import {getValues, saveValues} from "../../core/MemoryService";
+import BalanceService from "../BalanceService";
 
 let NameForm = () => {
   return (
     <div>
       
       <h1>MainPage</h1>
-      <StatedBalanceForm getValues={getValues} saveValues={saveValues}/>
+      <StatedBalanceForm balanceService={BalanceService.instance}/>
     </div>
   );
 }
